+++ FeedBackApp.Backend/Application/DTOs/Evaluation/UpdateQuestionnaireDTO.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace Application.DTOs.Questionnaire
+namespace Application.DTOs.Evaluation
 {
     public class UpdateQuestionnaireDTO
     {
