+++ FeedBackApp.Backend/Application/DTOs/Evaluation/UpdateResponseDTO.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Application.DTOs.Questionnaire
+﻿namespace Application.DTOs.Evaluation
 {
     public class UpdateResponseDTO
     {
