+++ FeedBackApp.Backend/Application/Services/Interfaces/IEvaluationService.cs
@@ -1,6 +1,9 @@
-﻿namespace Application.Services.Interfaces
+﻿using Application.DTOs.Evaluation;
+
+namespace Application.Services.Interfaces
 {
     public interface IEvaluationService
     {
+        public Task<UpdateResponseDTO> UpdateQuestionnaire(string id, UpdateQuestionnaireDTO dto);
     }
 }
