+++ FeedBackApp.Backend/Application/Validation/UpdateValidation/UpdateQuestionnaireValidator.cs
@@ -1,4 +1,4 @@
-﻿using Application.DTOs.Questionnaire;
+﻿using Application.DTOs.Evaluation;
 using FeedBackApp.Core.Model;
 using FluentValidation;
 using System.Collections.Generic;
