+++ FeedBackApp.Backend/FeedBackApp.Backend.Infrastructure/Persistence/Repository/QuestionnaireRepository.cs
@@ -103,36 +103,5 @@ namespace FeedBackApp.Backend.Infrastructure.Persistence.Repository
             await _context.SaveChangesAsync();
             return true;
         }
-
-        public async Task<bool> UpdateQuestionnaire(Questionnaire newQuestionnaire, Questionnaire oldQuestionnaire)
-        {
-            foreach(var oldAnswer in oldQuestionnaire.QuestionnaireResults)
-            {
-                var newAnswer = newQuestionnaire.QuestionnaireResults.FirstOrDefault(x => x.QuestionId == oldAnswer.QuestionId);
-                
-                if(newAnswer != null)
-                {
-                    oldAnswer.Answer = newAnswer.Answer;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            await _context.SaveChangesAsync();
-            return true;
-        }
-
-        public async Task<Questionnaire> GetQuestionnaireByIdAsync(string id)
-        {
-            return await _context.Questionnaires.FindAsync(id);
-        }
-
-        public async Task<QuestionnaireTemplate> GetQuestionTemplateBySurveyIdAsync(string surveyId)
-        {
-            string id = $"questiontemplates_{surveyId}";
-            return await _context.QuestionnaireTemplates.FindAsync(id);
-        }
     }
 }
