+++ FeedBackApp.Backend/FeedBackApp.Core/Repositories/IEvaluationRepository.cs
@@ -1,0 +1,12 @@
+﻿
+using FeedBackApp.Core.Model;
+
+namespace FeedBackApp.Core.Repositories
+{
+    public interface IEvaluationRepository
+    {
+        Task<bool> UpdateQuestionnaire(Questionnaire newQuestionnaire, Questionnaire oldQuestionnaire);
+        Task<Questionnaire> GetQuestionnaireByIdAsync(string id);
+        Task<QuestionnaireTemplate> GetQuestionTemplateBySurveyIdAsync(string surevyId);
+    }
+}
