+++ FeedBackApp.Backend/FeedBackApp.Core/Repositories/IQuestionnaireRepository.cs
@@ -9,8 +9,6 @@ namespace FeedBackApp.Core.Repositories
         Task<bool> DeleteSurveyMetadataAsync(Guid id);
         Task<bool> DeleteQuestionnairesBySurveyIdAsync(Guid surveyId);
         Task<bool> DeleteQuestionTemplateBySurveyIdAsync(Guid surveyId);
-        Task<bool> UpdateQuestionnaire(Questionnaire newQuestionnaire, Questionnaire oldQuestionnaire);
-        Task<Questionnaire> GetQuestionnaireByIdAsync(string id);
-        Task<QuestionnaireTemplate> GetQuestionTemplateBySurveyIdAsync(string surevyId);
+        
     }
 }
