--- FeedBackApp.Backend/FeedBackApp.Core/metadata.json
@@ -1,60 +1,0 @@
-{
-  "startDate": "2025-08-21",
-  "endDate": "2025-09-01",
-
-  "studentSets": [
-    {
-      "setId": "XI. A",
-      "studentEmails": [ "a@gimi.ro", "b@gimi.ro" ]
-    },
-    {
-      "setId": "XI. B",
-      "studentEmails": [ "c@gimi.ro", "d@gimi.ro" ]
-    }
-  ],
-
-  "questionnaireTemplate": [
-    {
-      "question": "How satisfied are you with the course?",
-      "type": "LikertScaleOneToFive"
-    },
-    {
-      "question": "How satisfied are you with the course?",
-      "type": "MultinomialSingleChoicee",
-      "answerOptions" :["igen", "nem", ""]
-    },
-    {
-      "question": "How satisfied are you with the course?",
-      "type": "MultipleChoice",
-      "answerOptions" :["igen", "nem", "nem nyilatkozok"]
-    },
-    {
-      "question": "How satisfied are you with the course?",
-      "type": "OpenEnded"
-    }
-  ],
-
-  "teachers": [
-    {
-      "email": "kovacs.maria@gimi.ro",
-      "name": "Kovács Mária"
-    },
-    {
-      "email": "kovacs.maria2@gimi.ro",
-      "name": "Kovács Mária2"
-    }
-  ],
-
-  "questionnaireCreationParams": [
-    {
-      "teacherEmail": "kovacs.maria@gimi.ro",
-      "subjectName": "Matematika",
-      "studentSetIds": [ "XI. A", "XI. B" ]
-    },
-    {
-      "teacherEmail": "kovacs.maria2@gimi.ro",
-      "subjectName": "Nemet",
-      "studentSetIds": [ "XI. B" ]
-    }
-  ]
-}
